Refactor: 게임 세부 설정 확인 이벤트 리스너를 각 페이지에서 설정하도록 변경
