+++ FE/src/pages/local/gameSetting/GameSettingDetailed.js
@@ -1,4 +1,5 @@
 import GameSettingDetailedComponent from '../../../components/gameSettingDetailed/GameSettingDetailedComponent.js';
+import { changeUrlData } from '../../../index.js';
 
 class GameSettingDetailed {
 	constructor() {
@@ -6,11 +7,27 @@ class GameSettingDetailed {
 	}
 
 	template(initial) {
-		return this.setting.template(initial, 'local');
+		this.battleMode = initial.battle_mode;
+		return this.setting.template(initial);
 	}
 
 	addEventListeners() {
 		this.setting.addEventListeners();
+		this.addConfirmEventListener();
+	}
+
+	addConfirmEventListener() {
+		const confirmButton = document.querySelector(
+			'.horizontalButton button:nth-child(2)'
+		);
+		confirmButton.addEventListener('click', () => {
+			const setting = {
+				1: 'gameSetting',
+				2: 'gameSettingTournament'
+			};
+
+			changeUrlData(setting[this.battleMode], this.setting.getModifiedData());
+		});
 	}
 }
 export default new GameSettingDetailed();
