+++ FE/src/pages/online/gameSetting/OnlineGameSettingDetailed.js
@@ -1,4 +1,5 @@
 import GameSettingDetailedComponent from '../../../components/gameSettingDetailed/GameSettingDetailedComponent.js';
+import { changeUrlData } from '../../../index.js';
 
 class OnlineGameSettingDetailed {
 	constructor() {
@@ -6,11 +7,27 @@ class OnlineGameSettingDetailed {
 	}
 
 	template(initial) {
-		return this.setting.template(initial, 'online');
+		this.battleMode = initial.battle_mode;
+		return this.setting.template(initial);
 	}
 
 	addEventListeners() {
 		this.setting.addEventListeners();
+		this.addConfirmEventListener();
+	}
+
+	addConfirmEventListener() {
+		const confirmButton = document.querySelector(
+			'.horizontalButton button:nth-child(2)'
+		);
+		confirmButton.addEventListener('click', () => {
+			const setting = {
+				1: 'onlineSetting',
+				2: 'onlineSettingTournament'
+			};
+
+			changeUrlData(setting[this.battleMode], this.setting.getModifiedData());
+		});
 	}
 }
 export default new OnlineGameSettingDetailed();
